moved functions to utils file
