+++ frontend/src/components/SingleListingPage/index.js
@@ -1,4 +1,4 @@
-import { useEffect } from "react";
+import { useEffect, useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { useParams } from "react-router-dom";
 import { clearListingStateAction, getSingleListingThunk } from "../../store/listings";
@@ -12,6 +12,10 @@ import SingleListingTitle from "./SingleListingTitle";
 export default function SingleListingPage() {
     const {listingId} = useParams()
     const dispatch = useDispatch()
+
+    const [showCreateReviewModal, setShowCreateReviewModal] = useState(false)
+
+
     const currentUser = useSelector(state => state.session.user)
     const allListings = useSelector(state => state.listings)
     const reviews = useSelector(state => state.reviews)
@@ -40,10 +44,18 @@ export default function SingleListingPage() {
                     <HostAndSpecs listing={listing}/>
                 </div>
                 <div className="createReviewBtnContainer">
-                    <ReviewFormModal currentUser={currentUser} listingId={listingId}/>
+                    <ReviewFormModal
+                    showCreateReviewModal={showCreateReviewModal}
+                    setShowCreateReviewModal={setShowCreateReviewModal}
+                    currentUser={currentUser}
+                    listingId={listingId}/>
                 </div>
                 <div className="reviewContainer">
-                    <ReviewsContainer reviews={reviews} currentUser={currentUser}/>
+                    <ReviewsContainer
+                    showCreateReviewModal={showCreateReviewModal}
+                    setShowCreateReviewModal={setShowCreateReviewModal}
+                    reviews={reviews}
+                    currentUser={currentUser}/>
                 </div>
 
             </div>
